Dwt 1342  remove draft references and updated PAT pages for new CO2 rule (#173)
* Update print statement from 'Hello' to 'Goodbye'

* Relplaced  Draft reference with Private Beta

Updated the API specification to reflect the current Private Beta status and revised the last updated date.

* Fix formatting issues inserted space before ##

* Updated and clarified information on OAuth

* Update FAQ section for Receipt of Waste API

* Updated title and CO2 Gherkin description

Updated the documentation for API Production Approval Tests, including the scenarios and requirements for waste receipt submissions.

* Updated link for api-landing-page in text

Updated the roadmap for the Receipt of Waste API, including recent releases and next steps for developers and Waste Receivers.

* Fix formatting issues and typos in roadmap documentation

* Update production approval tests documentation

* Revise production approval tests documentation CO2

Updated the documentation for the Receipt of Waste API Production Approval Tests, including detailed scenarios and technical exemptions.

* Format and update converted to bold rather than new 2 nd level heading